+++ unique_toolkit/CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.17.2] - 2025-10-23
-- Adding model `AZURE_GPT_5_PRO_2025_1006` to `language_model/info.py`
+- Adding model `AZURE_GPT_5_PRO_2025_1006` and `litellm:openai-gpt-5-pro` to `language_model/info.py`
 
 ## [1.17.1] - 2025-10-23
 - Fix hallucination check input with all cited reference chunks.